+++ src/main/kotlin/dev/ekvedaras/laravelquery/reference/ColumnPsiReference.kt
@@ -4,6 +4,9 @@ import com.intellij.database.util.DbUtil
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReferenceBase
 import dev.ekvedaras.laravelquery.models.DbReferenceExpression
+import dev.ekvedaras.laravelquery.services.LaravelQuerySettings
+import dev.ekvedaras.laravelquery.utils.DatabaseUtils.Companion.nameWithoutPrefix
+import dev.ekvedaras.laravelquery.utils.DatabaseUtils.Companion.tableNameWithoutPrefix
 
 class ColumnPsiReference(element: PsiElement) : PsiReferenceBase<PsiElement>(element) {
     override fun resolve(): PsiElement? {
@@ -12,8 +15,12 @@ class ColumnPsiReference(element: PsiElement) : PsiReferenceBase<PsiElement>(ele
 
         rangeInElement = target.ranges.last()
 
-        DbUtil.getDataSources(element.project).forEach { dataSource ->
-            val dbColumn = dataSource.findElement(target.column.find { tables.contains(it.tableName) })
+        DbUtil.getDataSources(element.project).filter {
+            LaravelQuerySettings.getInstance(element.project).interestedIn(it)
+        }.forEach { dataSource ->
+            val dbColumn = dataSource.findElement(target.column.find {
+                tables.contains(it.tableNameWithoutPrefix(element.project))
+            })
             if (dbColumn != null) {
                 return dbColumn
             }