+++ src/main/kotlin/dev/ekvedaras/laravelquery/reference/SchemaPsiReference.kt
@@ -4,6 +4,7 @@ import com.intellij.database.util.DbUtil
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReferenceBase
 import dev.ekvedaras.laravelquery.models.DbReferenceExpression
+import dev.ekvedaras.laravelquery.services.LaravelQuerySettings
 
 class SchemaPsiReference(element: PsiElement, private val type: DbReferenceExpression.Companion.Type) :
     PsiReferenceBase<PsiElement>(element) {
@@ -12,7 +13,9 @@ class SchemaPsiReference(element: PsiElement, private val type: DbReferenceExpre
 
         rangeInElement = target.ranges.first()
 
-        DbUtil.getDataSources(element.project).forEach { dataSource ->
+        DbUtil.getDataSources(element.project).filter {
+            LaravelQuerySettings.getInstance(element.project).interestedIn(it)
+        }.forEach { dataSource ->
             val dbSchema = dataSource.findElement(target.schema.firstOrNull())
             if (dbSchema != null) {
                 return dbSchema